remove unused code and rename
