+++ packages/typescript-plugin/src/ngHelperTsService/methodMapping.ts
