+++ packages/typescript-plugin/src/ngHelperTsService/utils.ts
